refactor: extract generic scheduler Caller interface from httpsched package into scheduler/calls package
