+++ httpcli/httpsched/httpsched.go
@@ -12,6 +12,7 @@ import (
 	"github.com/mesos/mesos-go/encoding"
 	"github.com/mesos/mesos-go/httpcli"
 	"github.com/mesos/mesos-go/scheduler"
+	"github.com/mesos/mesos-go/scheduler/calls"
 )
 
 const (
@@ -45,28 +46,11 @@ type (
 
 	// Caller is the public interface this framework scheduler's should consume
 	Caller interface {
+		calls.Caller
 		// httpDo is intentionally package-private; clients of this package may extend a Caller
 		// generated by this package by overriding the Call func but may not customize httpDo.
 		httpDo(encoding.Marshaler, ...httpcli.RequestOpt) (mesos.Response, error)
-
-		// Call issues a call to Mesos and properly manages call-specific HTTP response headers & data.
-		// Subscriptions that succeed return a non-nil Caller that should be used for duration of the subscription.
-		Call(*scheduler.Call) (mesos.Response, Caller, error)
 	}
-
-	// CallerAdapter is a convenience type that allows clients to more simply decorate an existing Caller.
-	CallerAdapter struct {
-		// Caller is expected to be decorated by a CallFunc implementation. This SHOULD NOT be a nil reference.
-		Caller
-		// CallFunc is invoked upon calls to CallerAdapter.Call; if not specified then the invocation is
-		// passed through to the embedded Caller.
-		CallFunc func(*scheduler.Call) (mesos.Response, Caller, error)
-	}
-
-	// Decorator funcs usually return a Caller whose behavior has been somehow modified
-	Decorator func(Caller) Caller
-
-	Decorators []Decorator
 
 	callerInternal interface {
 		Caller
@@ -86,24 +70,6 @@ type (
 	}
 )
 
-// Apply is a convenient, nil-safe applicator that returns the result of d(c) iff d != nil; otherwise c
-func (d Decorator) Apply(c Caller) (result Caller) {
-	if d != nil {
-		result = d(c)
-	} else {
-		result = c
-	}
-	return
-}
-
-func (ca *CallerAdapter) Call(call *scheduler.Call) (mesos.Response, Caller, error) {
-	f := ca.CallFunc
-	if f == nil && ca.Caller != nil {
-		f = ca.Caller.Call
-	}
-	return f(call)
-}
-
 func (ct *callerTemporary) httpDo(m encoding.Marshaler, opt ...httpcli.RequestOpt) (resp mesos.Response, err error) {
 	ct.callerInternal.WithTemporary(ct.opt, func() error {
 		if len(opt) == 0 {
@@ -117,7 +83,7 @@ func (ct *callerTemporary) httpDo(m encoding.Marshaler, opt ...httpcli.RequestOp
 	return
 }
 
-func (ct *callerTemporary) Call(call *scheduler.Call) (resp mesos.Response, caller Caller, err error) {
+func (ct *callerTemporary) Call(call *scheduler.Call) (resp mesos.Response, caller calls.Caller, err error) {
 	ct.callerInternal.WithTemporary(ct.opt, func() error {
 		resp, caller, err = ct.callerInternal.Call(call)
 		return nil
@@ -181,18 +147,8 @@ func (cli *client) httpDo(m encoding.Marshaler, opt ...httpcli.RequestOpt) (resp
 	}
 }
 
-// CallNoData is a convenience func that executes the given Call using the provided Caller
-// and always drops the response data.
-func CallNoData(caller Caller, call *scheduler.Call) error {
-	resp, _, err := caller.Call(call)
-	if resp != nil {
-		resp.Close()
-	}
-	return err
-}
-
 // Call implements Client
-func (cli *client) Call(call *scheduler.Call) (resp mesos.Response, exportCaller Caller, err error) {
+func (cli *client) Call(call *scheduler.Call) (resp mesos.Response, exportCaller calls.Caller, err error) {
 	var (
 		exec   = callerExec(prepareCall)
 		caller = callerInternal(cli)
@@ -332,50 +288,4 @@ func buildNewEndpoint(location, currentEndpoint string) (string, bool) {
 	}
 	current.Host = hostport.Host
 	return current.String(), true
-}
-
-var noopDecorator = Decorator(func(h Caller) Caller { return h })
-
-// If returns the receiving Decorator if the given bool is true; otherwise returns a no-op
-// Decorator instance.
-func (d Decorator) If(b bool) Decorator {
-	if d == nil {
-		return noopDecorator
-	}
-	result := noopDecorator
-	if b {
-		result = d
-	}
-	return result
-}
-
-// Apply applies the Decorators in the order they're listed such that the last Decorator invoked
-// generates the final (wrapping) Caller that is ultimately returned.
-func (ds Decorators) Combine() (result Decorator) {
-	actual := make(Decorators, 0, len(ds))
-	for _, d := range ds {
-		if d != nil {
-			actual = append(actual, d)
-		}
-	}
-	if len(actual) == 0 {
-		result = noopDecorator
-	} else {
-		result = Decorator(func(h Caller) Caller {
-			for _, d := range actual {
-				h = d(h)
-			}
-			return h
-		})
-	}
-	return
-}
-
-// CallerTracker invokes the tracking function `f` upon decorator invocation. This is a convenience
-// decorator that allows a client to receive callbacks when, in some context, a Caller is updated.
-func CallerTracker(f func(Caller)) Decorator {
-	return func(c Caller) Caller {
-		f(c)
-		return c
-	}
 }
